refactor(find-and-replace): reorganize source files into subdirectories and split modules
