+++ packages/extension-find-and-replace/src/index.ts
@@ -1,8 +1,8 @@
 import { FindAndReplace } from './find-and-replace.js'
 
 export * from './find-and-replace.js'
-export * from './plugin.js'
-export * from './search.js'
+export * from './plugin/plugin.js'
+export * from './search/search.js'
 export * from './types.js'
 
 export default FindAndReplace
