+++ packages/extension-find-and-replace/src/types.ts
@@ -1,4 +1,4 @@
-import type { SearchResult } from './search.js'
+import type { SearchResult } from './search/search.js'
 
 export interface FindAndReplaceOptions {
   /**
