+++ packages/extension-find-and-replace/src/utils/createDecoration.ts
@@ -1,7 +1,7 @@
 import { Decoration } from '@tiptap/pm/view'
 
-import { currentResultClass, resultClass } from '../constants.js'
-import type { SearchResult } from '../search.js'
+import { currentResultClass, resultClass } from '../constants/constants.js'
+import type { SearchResult } from '../search/search.js'
 
 export function createDecoration(
   result: SearchResult,
