+++ packages/extension-find-and-replace/src/utils/createDecorations.ts
@@ -1,7 +1,7 @@
 import type { Node } from '@tiptap/pm/model'
 import { DecorationSet } from '@tiptap/pm/view'
 
-import type { SearchResult } from '../search.js'
+import type { SearchResult } from '../search/search.js'
 
 import { createDecoration } from './createDecoration.js'
