+++ packages/extension-find-and-replace/src/utils/createState.ts
@@ -1,8 +1,8 @@
 import type { Node } from '@tiptap/pm/model'
 
 import type { FindAndReplaceOptions } from '../types.js'
-import type { FindAndReplacePluginState } from '../plugin-state.js'
-import { searchDocument } from '../search.js'
+import type { FindAndReplacePluginState } from '../plugin/plugin-state.js'
+import { searchDocument } from '../search/search.js'
 
 import { createDecorations } from './createDecorations.js'
