+++ packages/extension-find-and-replace/src/utils/getTextNodeResult.ts
@@ -1,6 +1,6 @@
 import type { EditorState } from '@tiptap/pm/state'
 
-import type { SearchResult } from '../search.js'
+import type { SearchResult } from '../search/search.js'
 
 export interface TextNodeResult {
   from: number
