+++ packages/extension-find-and-replace/src/utils/hasCurrentIndex.ts
@@ -1,4 +1,4 @@
-import type { FindAndReplaceMeta } from '../plugin-state.js'
+import type { FindAndReplaceMeta } from '../plugin/plugin-state.js'
 
 export function hasCurrentIndex(meta: FindAndReplaceMeta | undefined): boolean {
   return !!meta && 'currentIndex' in meta
