+++ packages/extension-find-and-replace/src/utils/isCurrentResult.ts
@@ -1,4 +1,4 @@
-import type { SearchResult } from '../search.js'
+import type { SearchResult } from '../search/search.js'
 
 import { sameResult } from './sameResult.js'
