+++ packages/extension-find-and-replace/src/utils/isNewSearch.ts
@@ -1,4 +1,4 @@
-import type { FindAndReplaceMeta } from '../plugin-state.js'
+import type { FindAndReplaceMeta } from '../plugin/plugin-state.js'
 
 export function isNewSearch(meta: FindAndReplaceMeta | undefined): boolean {
   return !!meta && 'searchTerm' in meta
