+++ packages/extension-find-and-replace/src/utils/replaceAllResults.ts
@@ -1,6 +1,6 @@
 import type { EditorState, Transaction } from '@tiptap/pm/state'
 
-import type { SearchResult } from '../search.js'
+import type { SearchResult } from '../search/search.js'
 import { groupResults } from './groupResults.js'
 import { replaceGroup } from './replaceGroup.js'
