+++ packages/extension-find-and-replace/src/utils/replaceGroup.ts
@@ -1,6 +1,6 @@
 import type { Transaction } from '@tiptap/pm/state'
 
-import type { SearchResult } from '../search.js'
+import type { SearchResult } from '../search/search.js'
 import type { ResultGroup } from './groupResults.js'
 
 function replaceSeparateResults(
