+++ packages/extension-find-and-replace/src/utils/resultInTextblocks.ts
@@ -1,4 +1,4 @@
-import type { SearchResult } from '../search.js'
+import type { SearchResult } from '../search/search.js'
 
 import type { TextblockRange } from './types.js'
