+++ packages/extension-find-and-replace/src/utils/sameResult.ts
@@ -1,4 +1,4 @@
-import type { SearchResult } from '../search.js'
+import type { SearchResult } from '../search/search.js'
 
 export function sameResult(first: SearchResult, second: SearchResult): boolean {
   return first.from === second.from && first.to === second.to
